changed: default share card
