+++ templates/mdnext-blog/pages/blog/[...slug].js
@@ -21,7 +21,7 @@ export default function BlogPost({ mdxSource, frontMatter }) {
 
   return (
     <Layout
-      title={title}
+      title={title ? title : "MDNext Blog Starter"}
       description={description}
       openGraph={{
         title,
